Ability to copy files.
Invoke-CopyActions can copy files from different sources to destinations. It is possible to use environment variables in DOS format (e.g. %CommonProgramFiles%\test.txt) and if the path is not an absolute (i.e. routed) path, it will assume that the path is somewhere under the folder which contains the script.